+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## 0.0.8 - 2025-09-17
+
+### Added
+
+ * Support for Illumina NextSeq 2000 and MiSeq i100 Plus runs ([#145])
+ * New family of `Analysis` classes for Illumina Analysis directories to
+   supersede the old `Alignment` class ([#145])
+ * `instrument_type` property for `Run` class to identify the sequencer model
+   of a given run ([#143])
+ * Illumina's "v2" sample sheets, used by newer instruments like the MiSeq i100
+   Plus and NextSeq 200, are now supported ([#142])
+
+### Changed
+
+ * Python from 3.9 to 3.11 is supported ([#140])
+
+[#145]: https://github.com/ShawHahnLab/umbra/pull/145
+[#143]: https://github.com/ShawHahnLab/umbra/pull/143
+[#142]: https://github.com/ShawHahnLab/umbra/pull/142
+[#140]: https://github.com/ShawHahnLab/umbra/pull/140
+
 ## 0.0.7 - 2025-07-11
 
 ### Added